backend/requirements.txt
Django==5.0.*
djangorestframework==3.15.*
psycopg2-binary
python-decouple
django-cors-headers
django-filter
Pillow
gunicorn
